Refactored interfaces and repositories

+++ Models/CategoryRepository.cs
@@ -5,34 +5,39 @@ using System.Threading.Tasks;
 
 namespace eCommerce.Models
 {
-    public class CategoryRepository : ICategoryRepository
+    public class CategoryRepository : IDataRepository<Category, Category>
     {
         private readonly DataContext context;
         public CategoryRepository(DataContext dataContext) 
             => context = dataContext;
+        
 
-        public void AddCategory(Category category)
+        public Category CreateItem(Category item)
         {
-            context.Categories.Add(category);
+            context.Categories.Add(item);
             context.SaveChanges();
+            return item;
+        }     
+
+        public Category DeleteItem(long id)
+        {
+            Category category = FindItemById(id);
+            if (category != null)
+            {
+                context.Categories.Remove(category);
+                context.SaveChanges();
+            }
+            return category;
+        }            
+
+        public Category EditItem(Category item)
+        {
+            context.Categories.Update(item);
+            context.SaveChanges();
+            return item;
         }
 
-        public void DeleteCategory(Category category)
-        {
-            context.Categories.Remove(category);
-            context.SaveChanges();
-        }
-
-        public void EditCategory(Category category)
-        {
-            context.Categories.Update(category);
-            context.SaveChanges();
-        }
-
-        public Category FindCategoryById(long id)
-            => context.Categories.FirstOrDefault(r => r.CategoryId == id);        
-
-        public IEnumerable<Category> GetCategories() 
-            => context.Categories.ToArray();       
+        public Category FindItemById(long id) => context.Categories.FirstOrDefault(r => r.CategoryId == id); 
+        public IEnumerable<Category> GetAllItems() => context.Categories.ToArray();            
     }
 }
